Help me fix if you can
I want to create an object, Radio. Radio is the son of a parent class, Serial.

The comments are if I don't do OOP (I still am using objects but not coding in OOP).

The methods of the object don't work, even though it is exactly as it is in the comments.